ADDED BREADCRUMB, BANNER, SIDEBAR, TOPBAR
Added an automatic Breadcrumb for easy navigation accross pages and edited some routes to easily
integrate with the breadcrumb

Added Mobile responsive topbar that contains links to the homepage and
about us page

Added the banner to the guest layout. Added banner upload to the site
settings to allow admin upload a banner from the admin dashboard

Added a sidebar to contain contact information.

Added animations to the menu pages